+++ tests/Spreads.Collections.Tests/SortedDequeTests.cs
@@ -30,15 +30,21 @@ namespace Spreads.Collections.Tests {
             var fst = sd.RemoveFirst();
             Assert.AreEqual(sd.First, 3);
             sd.Add(4);
-            Assert.AreEqual(1, sd.IndexOfElement(4));
-            Assert.AreEqual(2, sd.IndexOfElement(5));
-            Assert.AreEqual(3, sd.IndexOfElement(7));
+            Assert.AreEqual(1, sd.ToList().IndexOf(4));
+            Assert.AreEqual(2, sd.ToList().IndexOf(5));
+            Assert.AreEqual(3, sd.ToList().IndexOf(7));
 
             var last = sd.RemoveLast();
             sd.Add(8);
-            Assert.AreEqual(1, sd.IndexOfElement(4));
-            Assert.AreEqual(2, sd.IndexOfElement(5));
-            Assert.AreEqual(3, sd.IndexOfElement(8));
+            Assert.AreEqual(1, sd.ToList().IndexOf(4));
+            Assert.AreEqual(2, sd.ToList().IndexOf(5));
+            Assert.AreEqual(3, sd.ToList().IndexOf(8));
+
+            sd.Add(6);
+            Assert.AreEqual(1, sd.ToList().IndexOf(4));
+            Assert.AreEqual(2, sd.ToList().IndexOf(5));
+            Assert.AreEqual(3, sd.ToList().IndexOf(6));
+            Assert.AreEqual(4, sd.ToList().IndexOf(8));
         }
 
         [Test]
@@ -51,7 +57,39 @@ namespace Spreads.Collections.Tests {
             }
 
             for (int i = 0; i < 4; i++) {
-                Assert.AreEqual(i, sd.IndexOfElement(i));
+                Assert.AreEqual(i, sd.ToList().IndexOf(i));
+            }
+        }
+
+
+        [Test]
+        public void CouldAddRemoveWithFixedSize() {
+            var sd = new SortedDeque<int>();
+            for (int i = 0; i < 4; i++) {
+                sd.Add(i);
+            }
+
+            for (int i = 4; i < 100; i++)
+            {
+                sd.RemoveFirst();
+                sd.Add(i);
+                Assert.AreEqual(i-3, sd.First);
+                Assert.AreEqual(i, sd.Last);
+            }
+        }
+
+        [Test]
+        public void CouldAddRemoveIncreasing() {
+            var sd = new SortedDeque<int>();
+            for (int i = 0; i < 4; i++) {
+                sd.Add(i);
+            }
+
+            for (int i = 4; i < 100; i++) {
+                if(i % 2 == 0) sd.RemoveFirst();
+                sd.Add(i);
+                //Assert.AreEqual(i - 3, sd.First);
+                Assert.AreEqual(i, sd.Last);
             }
         }