+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 0.0.47 - August 13, 2015
+* Implement SortedDeque, which is much faster than FixedMinHeap or SortedList1 (TODO test off-by-one bugs for all cases in insert)
+* Implement FixedMinHeap and SortedList1
+
 #### 0.0.46 - August 11, 2015
 * ZipNCursor MoveNext single-threaded implementation